requirements.txt
colorama==0.3.7
Office365_REST_Python_Client==2.3.2
PyYAML==5.4.1
tqdm==4.55.1
